+++ lib/better_player.dart
@@ -15,3 +15,4 @@ export 'src/list/better_player_list_video_player.dart';
 export 'src/list/better_player_list_video_player_controller.dart';
 export 'src/subtitles/better_player_subtitles_source.dart';
 export 'src/subtitles/better_player_subtitles_source_type.dart';
+export 'src/controls/better_player_overflow_menu_item.dart';